Re-organize store + redux-toolkit + redux chrome extension
