+++ package.json
@@ -32,9 +32,11 @@
     "package:checksums": "bash scripts/checksum.sh"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.5.0",
     "bluebird": "3.7.2",
     "chardet": "1.3.0",
     "classnames": "2.2.6",
+    "electron-devtools-installer": "^3.1.1",
     "font-awesome": "4.7.0",
     "globby": "11.0.1",
     "history": "4.10.1",
@@ -60,7 +62,6 @@
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
     "react-simple-input": "0.4.2",
-    "redux": "4.0.5",
     "redux-persist": "^6.0.0",
     "redux-thunk": "2.3.0",
     "semver": "7.3.4",
@@ -70,6 +71,7 @@
   "devDependencies": {
     "@types/bluebird": "3.5.33",
     "@types/classnames": "2.2.11",
+    "@types/electron-devtools-installer": "^2.2.0",
     "@types/history": "4.7.8",
     "@types/lodash-es": "4.17.4",
     "@types/mime-types": "2.1.0",
