+++ src/renderer/components/Header/Header.tsx
@@ -6,9 +6,9 @@ import KeyBinding from 'react-keybinding-component';
 import PlayingBar from '../PlayingBar/PlayingBar';
 import PlayerControls from '../PlayerControls/PlayerControls';
 
-import * as LibraryActions from '../../actions/LibraryActions';
+import * as LibraryActions from '../../store/actions/LibraryActions';
 import { isCtrlKey } from '../../lib/utils-platform';
-import { RootState } from '../../reducers';
+import { RootState } from '../../store/reducers';
 import { TrackModel, PlayerStatus, Repeat } from '../../../shared/types/museeks';
 
 import styles from './Header.module.css';
