+++ src/renderer/components/PlayerControls/PlayerControls.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Icon from 'react-fontawesome';
 import VolumeControl from '../VolumeControl/VolumeControl';
 
-import * as PlayerActions from '../../actions/PlayerActions';
+import * as PlayerActions from '../../store/actions/PlayerActions';
 import { PlayerStatus } from '../../../shared/types/museeks';
 
 import styles from './PlayerControls.module.css';
