+++ src/renderer/components/QueueList/QueueList.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import QueueListItem from '../QueueListItem/QueueListItem';
 
-import * as QueueActions from '../../actions/QueueActions';
+import * as QueueActions from '../../store/actions/QueueActions';
 
 import { getStatus } from '../../lib/utils-library';
 import { TrackModel } from '../../../shared/types/museeks';
