+++ src/renderer/main.tsx
@@ -10,7 +10,7 @@ import { Provider } from 'react-redux';
 
 import Root from './Root';
 import Router from './router/Router';
-import store from './store';
+import store from './store/store';
 
 /*
 |--------------------------------------------------------------------------
