+++ src/renderer/store/actions/AppActions.ts
@@ -1,10 +1,10 @@
 import os from 'os';
 import electron from 'electron';
 
-import Player from '../lib/player';
-import { browserWindows, config } from '../lib/app';
-import * as utils from '../lib/utils';
-import * as coverUtils from '../../shared/lib/utils-cover';
+import Player from '../../lib/player';
+import { browserWindows, config } from '../../lib/app';
+import * as utils from '../../lib/utils';
+import * as coverUtils from '../../../shared/lib/utils-cover';
 
 import * as LibraryActions from './LibraryActions';
 import * as PlaylistsActions from './PlaylistsActions';
