+++ src/renderer/store/actions/LibraryActions.ts
@@ -7,13 +7,13 @@ import * as queue from 'queue';
 
 import store from '../store';
 
-import types from '../constants/action-types';
+import types from '../action-types';
 
-import * as app from '../lib/app';
-import * as utils from '../lib/utils';
-import * as m3u from '../lib/utils-m3u';
-import { SortBy, TrackModel } from '../../shared/types/museeks';
-import { SUPPORTED_PLAYLISTS_EXTENSIONS, SUPPORTED_TRACKS_EXTENSIONS } from '../../shared/constants';
+import * as app from '../../lib/app';
+import * as utils from '../../lib/utils';
+import * as m3u from '../../lib/utils-m3u';
+import { SortBy, TrackModel } from '../../../shared/types/museeks';
+import { SUPPORTED_PLAYLISTS_EXTENSIONS, SUPPORTED_TRACKS_EXTENSIONS } from '../../../shared/constants';
 import * as PlaylistsActions from './PlaylistsActions';
 import * as ToastsActions from './ToastsActions';
